- erased old ui - created new ui, based on components - created new project animation - moved float animation to animation project
Former-commit-id: d0105a2335b522ddc6475d4173b7a9563ae3a8d4
Former-commit-id: 034351700ce507af82c15c0afad7b1bd2643beba

+++ src/rendering/gBufferRenderPass.h
@@ -14,8 +14,8 @@ namespace phi
     {
     private:
         gl* _gl;
-        size_t _w;
-        size_t _h;
+        float _w;
+        float _h;
     public:
         shader* shader;
         framebuffer* framebuffer;
@@ -25,7 +25,7 @@ namespace phi
         void initShader();
         void initRenderTargets();
     public:
-        RENDERING_API gBufferRenderPass(gl* gl, size_t w, size_t h);
+        RENDERING_API gBufferRenderPass(gl* gl, float w, float h);
         RENDERING_API ~gBufferRenderPass();
 
         RENDERING_API void update();
